Rename top-level IO extension headers.
