+++ io/test/all_formats_test.cpp
@@ -9,12 +9,12 @@
 // This test is more of a compilation test.
 
 
-#include <boost/gil/extension/io/png_all.hpp>
-#include <boost/gil/extension/io/bmp_all.hpp>
-#include <boost/gil/extension/io/jpeg_all.hpp>
-#include <boost/gil/extension/io/pnm_all.hpp>
-#include <boost/gil/extension/io/targa_all.hpp>
-#include <boost/gil/extension/io/tiff_all.hpp>
+#include <boost/gil/extension/io/png.hpp>
+#include <boost/gil/extension/io/bmp.hpp>
+#include <boost/gil/extension/io/jpeg.hpp>
+#include <boost/gil/extension/io/pnm.hpp>
+#include <boost/gil/extension/io/targa.hpp>
+#include <boost/gil/extension/io/tiff.hpp>
 
 #define BOOST_TEST_MODULE all_formats_test
 #include <boost/test/unit_test.hpp>
