+++ io/test/bmp_read_test.cpp
@@ -10,7 +10,7 @@
 #include <boost/type_traits/is_same.hpp>
 
 #include <boost/gil.hpp>
-#include <boost/gil/extension/io/bmp_all.hpp>
+#include <boost/gil/extension/io/bmp.hpp>
 
 #include "scanline_read_test.hpp"
