+++ io/test/jpeg_test.cpp
@@ -15,7 +15,7 @@
 #include <boost/gil.hpp>
 
 #define BOOST_GIL_IO_ADD_FS_PATH_SUPPORT
-#include <boost/gil/extension/io/jpeg_all.hpp>
+#include <boost/gil/extension/io/jpeg.hpp>
 
 #include "mandel_view.hpp"
 #include "paths.hpp"
