+++ io/test/png_test.cpp
@@ -14,7 +14,7 @@
 
 //#define BOOST_GIL_IO_PNG_FLOATING_POINT_SUPPORTED
 //#define BOOST_GIL_IO_PNG_FIXED_POINT_SUPPORTED
-#include <boost/gil/extension/io/png_all.hpp>
+#include <boost/gil/extension/io/png.hpp>
 
 #include <boost/type_traits/is_same.hpp>
