+++ io/test/targa_test.cpp
@@ -13,7 +13,7 @@
 #include <fstream>
 
 #include <boost/gil.hpp>
-#include <boost/gil/extension/io/targa_all.hpp>
+#include <boost/gil/extension/io/targa.hpp>
 
 #include <boost/filesystem/convenience.hpp>
