+++ io/test/tiff_file_format_test.cpp
@@ -14,7 +14,7 @@
 //#define BOOST_TEST_MODULE tiff_file_format_test_module
 #include <boost/test/unit_test.hpp>
 
-#include <boost/gil/extension/io/tiff_all.hpp>
+#include <boost/gil/extension/io/tiff.hpp>
 
 #include "paths.hpp"
 #include "scanline_read_test.hpp"
