+++ io/test/tiff_subimage_test.cpp
@@ -8,7 +8,7 @@
 //#define BOOST_TEST_MODULE tiff_subimage_test_module
 #include <boost/test/unit_test.hpp>
 
-#include <boost/gil/extension/io/tiff_all.hpp>
+#include <boost/gil/extension/io/tiff.hpp>
 
 #include <fstream>
 #include <sstream>
