+++ io/test/tiff_write_test.cpp
@@ -11,7 +11,7 @@
 #include <boost/gil.hpp>
 
 #include <boost/gil/extension/io/detail/typedefs.hpp>
-#include <boost/gil/extension/io/tiff_all.hpp>
+#include <boost/gil/extension/io/tiff.hpp>
 
 #include "color_space_write_test.hpp"
 #include "mandel_view.hpp"
